fix: [#57426] timeout issues
We're experiencing timeout errors in various integrations, but it does
not look like we ever send the request to OS2mo/Keycloak/etc. This MR
bumps httpx to a patched version which in turn depends on patched
version of httpcore which includes
https://github.com/encode/httpcore/pull/802, which I suspect is our
issue.

Another related issue has also been fixed
(https://github.com/encode/httpcore/pull/641), and is already included
in httpcore 0.18.0.